Committed Changes made in the borland tree Committed Changes made by Mike Nordel re, 64 bit io for win32 and bug fixes

+++ jrd/pag.h
@@ -42,9 +42,9 @@ typedef struct pgc {
     SLONG	pgc_high_water;		/* Lowest PIP with space */
     SLONG	pgc_ppp;		/* Pages per pip */
     SLONG	pgc_pip;		/* First pointer page */
-    int		pgc_bytes;		/* Number of bytes of bit in PIP */
-    int		pgc_tpt;		/* Transactions per TIP */
-    int		pgc_gpg;		/* Generators per generator page */
+    ULONG	pgc_bytes;		/* Number of bytes of bit in PIP */
+    ULONG	pgc_tpt;		/* Transactions per TIP */
+    ULONG	pgc_gpg;		/* Generators per generator page */
 } *PGC;
 
 #endif /* _JRD_PAG_H_ */
